+++ index.js
@@ -1,0 +1,3 @@
+export {EtoolsUpload} from './EtoolsUpload.js';
+export {EtoolsUploadMulti} from './EtoolsUploadMulti.js';
+export {uploadFileStoredInDexie} from './upload-helper.js';
