+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@unicef-polymer/etools-upload",
-  "version": "3.3.1",
+  "version": "3.4.0-rc.1",
   "homepage": "https://github.com/unicef-polymer/etools-upload",
   "repository": {
     "type": "git",
@@ -9,6 +9,8 @@
   "description": "Upload files",
   "author": "The Polymer Authors",
   "license": "BSD-3-Clause",
+  "main": "index.js",
+  "module": "index.js",
   "scripts": {
     "lint:javascript": "eslint . --ext js,html && exit 0",
     "start": "polymer serve",
